Refactoring: extract methods about constructor codegen out
 #KT-25614 In Progress
